Added code that allows the output of all propensities at each step, which is useful for thinking about implementing tau leaping.
Created a fix for reading in 'delete' molecule operations, that did not properly handle the given xml ID.  This will have to be updated shortly, however, to handle the special delete molecule keywords in BNGL.

Also fixed a bug when removing molecules that disrupted how observables are counted.  Molecules where using an old function to update observable counts that did not work correctly.