Adds the ability for a single molecule agent to be mapped to the same reaction. Preparing DORRxnClass::tryToAdd to use it.
This commit changes the Molecule::rxnListMappingId data field in NFcore.hh to be a map of vectors instead of int*
This is to allow a single molecule agent to be mapped to more than one reaction.. Works still needs to be done
in the getRxnListMappingID method. (it only returns the first match right now). Additional work is also needed in setRxnListMapping
for when a mapping is removed (some refactoring to allow for a delete method might be desired).

Additionally it modified templateMolecule::compare method to allow for a memory of sorts where previous
matches can be remembered and skipped. This is to allow storing multiple matches per templateMolecule-molecule agent pair
Previously it would only return the last mapping found.

Additionally work was started in DORreaction to use this functionality: when adding mappingSets to the reactantTree for a DORReaction it
will now exhaustively set all possible mappings a molecule agent can have. Work is still needed to correctly address the fact that multiple
mapping id's can be used for reaction per agent though.

DORRxnClass::tryToAdd now