Adding code to handle removal of reaction-agent multi mapping
Added code to handle the update of reaction-molecule agent pairing removal.
Before it would only check that a reaction and molecule agent no longer map
to each other before removing the relationship completely. However now it must consider
the case where a molecule agent/reaction may stop mapping to each other in some ways but not
others. As such it must be selective in which mappings to remove.

TODO: the comparison function together with the new repeat detection system
is not handling correctly case where the napping does not involve any bonds,
leading to infinite loops.