several bugfixes for cases resulting in infinite loops
Fixed several cases that would lead the system to an infinite loop. These are related
to the checks in DORreaction where it must verify if a mapping has been found before adding
it to the system. With these changes blbr now runs to completion.

TODO: the necessary checks for previous mappings have resulted in a performance hit. Those need to be optimized.