Adding extensive mapping checking to elereactions. Also added a look-ahead check for stopping early
Added the same infrastructure we developed for extensive pattern mapping checking to normal reactions with elementary rates.

Additionally the singnature of the templateMolecules::compare method was changed from bool to pair<bool,bool>. The first bool
returns the same information the previous signature returned (a mapping was found). The second bool is used to indicate that
all possible mappings have been found and the method does not need to be called again.