Rename artifact and zip it.
Change from NFsim-x64.exe to NFsim-Winx64.exe.

Also, there seems to be a problem when I try to download the binary.
With FireFox, it thinks the executable is a zip file and will download
it as NFsim-x64.exe.zip for instance. And with chrome, it downloads it
as an executable, but the file won't run.