Traversal of MappingSet and not BondedNeigborhood
This code is buggy for TASEP since it ignores state changes of
trailing ribosomes that are stacked behind a leading ribosome but are
not bonded to it.
To solve this problem, I am going to look at the MappingSet of mRNA
molecules to identify the spatial region around the state change and
include any associated bonded molecules in the product list.
Will involve introducing notion of a polymer molecule and spatial
states.