Removed failing commands from osx_install.sh
If you looked at the travis logs, it looks like the only thing the
osx_install script was doing anything in the osx_install script was the
brew update, so I removed it and added that one line to travis.yml. It's
not even clear if that one line is necessary.