Connected and full versions identical first time
Based on my tests, it looks like I am inferring and processing the
connected reactions correctly.

I removed some debugging cruft.
TODO: Further optimizations that can/should be done:

1. Instead of iterating over all reactions for each molecule to update
its membership, iterate over all the connected reactions for the fired
Reaction in =MoleculeType::updateRxnMembership()=. This is a much
smaller set with the network topology of ribosomes and mRNAs.

2. Remove =onTheFlyObservable= calculation. Right now, this loop
needlessly slows down the fire. But it can be readily removed because it
is automatically assumed for the =TotalRate= calculation for
=initiation=.

3. Identify network connectivity within =NFsim= itself. This will speed
up the reaction considerably.