Corrected polymer search for ribosome collisions
In my previous code, when polymer molecules did not undergo changes,
then they were not added to the product list. Now they get added if they
are bonded to any of the non-polymeric molecules that change.

I also added code for outputting all the reactions that undergo change
in their rates during firing of each reaction. This is useful for
checking but should be skipped during actual simulations to speed things
up.