Gave CLI for polymer use and infer connectivity
I cleaned the new parts of the code and gave command line options for
turning on inferring network connetivity (works out of the box) and
using polymer property of molecules (needs modification of
PySB and BioNetGen to specify and parse polymer properties).

At this point, it is ready for simulation.